mysite/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==4.2.7
django-anymail==10.2
django-bootstrap-v5==1.0.11
django-bootstrap4==23.2
django-environ==0.11.2
django-filter==23.5
gunicorn==21.2.0
idna==3.6
packaging==23.2
pycparser==2.21
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
